// src/app/models/Participante.ts
export class Participante {    
    lugar: string;
    chip: string;
    dorsal: string;
    nombreParticipante: string;
    categoria: string;
    genero: string;
    tiempoOficial: string;
    tiempoChip: string;
    pace: string;
    status: string;
    evento: string;
    distancia: string;
    fecha: string;
}